Make changes to compile Ocotiger on MSVC 2017
* Fixed wrong header and source file names in CMakeLists.txt
* Make Silo a requirement in CMakeLists.txt
* Ignore more MSVC warnings in CMakeLists.txt
* Add cmake-build-debug to .gitignore
* Remove extra typename statements in src/common_kernel/kernel_simd_types.hpp
* Fixed variadic macro in src/options.hpp
* Replace asprintf with hpx::util::format
* Add missing Vc/Vc includes
* Move contents of sedovf.c C to sedovf_c.cpp
* Use boost::multiprecision::cpp_bin_float_quad instead of GCC's __float128 (N/A in MSVC) (src/test_problems/blast/sedov.cpp)
* std::sqrt is not constexpr in the standard or on MSVC. (src/test_problems/marshak/marshak.cpp)
* Make test1/validate.py Python 3 compatible
