+++ src/common_kernel/kernel_simd_types.hpp
@@ -1,6 +1,8 @@
 #pragma once
 // #include "taylor.hpp"
 
+#include <cstdint>
+
 #include <Vc/Vc>
 
 #ifdef __AVX__
@@ -10,13 +12,13 @@
 // using m2m_int_vector = Vc::datapar<int64_t, Vc::datapar_abi::avx512>;
 //#elif defined(Vc_HAVE_AVX)
 #endif
-#ifdef __AVX2__  // assumes avx 2
-using m2m_vector = typename Vc::Vector<double, Vc::VectorAbi::Avx>;
-using m2m_int_vector = typename Vc::Vector<int32_t, Vc::VectorAbi::Avx>;
+#ifdef __AVX2__  // assumes AVX 2
+using m2m_vector = Vc::Vector<double, Vc::VectorAbi::Avx>;
+using m2m_int_vector = Vc::Vector<int32_t, Vc::VectorAbi::Avx>;
 // using m2m_int_vector = typename Vc::datapar<int64_t, Vc::datapar_abi::avx>;
 #else                         // falling back to fixed_size types
-using m2m_vector = typename Vc::Vector<double, Vc::VectorAbi::Best<double>>;
-using m2m_int_vector = typename Vc::Vector<int32_t, Vc::VectorAbi::Best<int32_t>>;
+using m2m_vector = Vc::Vector<double>;
+using m2m_int_vector = Vc::Vector<std::int32_t>;
 #endif
 
 // using multipole_v = taylor<4, m2m_vector>;
