+++ src/test_problems/marshak/marshak.cpp
@@ -2,7 +2,7 @@
 
 /*----------------------------------------------------------------
  Laplace Inversion Source Code
- Copyright Â© 2010 James R. Craig, University of Waterloo
+ Copyright © 2010 James R. Craig, University of Waterloo
  ----------------------------------------------------------------*/
 
 #ifndef TESTME
@@ -139,7 +139,11 @@ double LaplaceInversion(const function<cmplex(const cmplex &s)>& F, const double
 
 static constexpr auto one = cmplex(1, 0);
 static constexpr double eps = 1.0;
+#ifdef _MSC_VER
+static constexpr auto root3 = cmplex(1.73205080757, 0.0);
+#else
 static constexpr auto root3 = cmplex(std::sqrt(3), 0.0);
+#endif
 static constexpr double theta_inc = 1.0;
 static constexpr double kappa = 1.0;
 static constexpr double c = 1.0;
